Refactor common parsing logic
