+++ src/crashlytics/addNote.ts
@@ -2,6 +2,7 @@ import { Client } from "../apiv2";
 import { logger } from "../logger";
 import { FirebaseError } from "../error";
 import { crashlyticsApiOrigin } from "../api";
+import { parseProjectNumber } from "./utils";
 
 const TIMEOUT = 10000;
 
@@ -15,12 +16,8 @@ type NoteRequest = {
 };
 
 export async function addNote(appId: string, issueId: string, note: string): Promise<string> {
+  const requestProjectId = parseProjectNumber(appId);
   try {
-    const requestProjectId = parseProjectId(appId);
-    if (requestProjectId === undefined) {
-      throw new FirebaseError("Unable to get the projectId from the AppId.");
-    }
-
     const response = await apiClient.request<NoteRequest, string>({
       method: "POST",
       headers: {
@@ -39,12 +36,4 @@ export async function addNote(appId: string, issueId: string, note: string): Pro
       { original: err },
     );
   }
-}
-
-function parseProjectId(appId: string): string | undefined {
-  const appIdParts = appId.split(":");
-  if (appIdParts.length > 1) {
-    return appIdParts[1];
-  }
-  return undefined;
 }
