+++ src/crashlytics/deleteNote.ts
@@ -2,6 +2,7 @@ import { Client } from "../apiv2";
 import { logger } from "../logger";
 import { FirebaseError } from "../error";
 import { crashlyticsApiOrigin } from "../api";
+import { parseProjectNumber } from "./utils";
 
 const TIMEOUT = 10000;
 
@@ -11,12 +12,8 @@ const apiClient = new Client({
 });
 
 export async function deleteNote(appId: string, issueId: string, noteId: string): Promise<void> {
+  const requestProjectId = parseProjectNumber(appId);
   try {
-    const requestProjectId = parseProjectId(appId);
-    if (requestProjectId === undefined) {
-      throw new FirebaseError("Unable to get the projectId from the AppId.");
-    }
-
     await apiClient.request<void, void>({
       method: "DELETE",
       path: `/projects/${requestProjectId}/apps/${appId}/issues/${issueId}/notes/${noteId}`,
@@ -29,12 +26,4 @@ export async function deleteNote(appId: string, issueId: string, noteId: string)
       { original: err },
     );
   }
-}
-
-function parseProjectId(appId: string): string | undefined {
-  const appIdParts = appId.split(":");
-  if (appIdParts.length > 1) {
-    return appIdParts[1];
-  }
-  return undefined;
 }
