+++ src/crashlytics/listTopVersions.ts
@@ -2,6 +2,7 @@ import { Client } from "../apiv2";
 import { logger } from "../logger";
 import { FirebaseError } from "../error";
 import { crashlyticsApiOrigin } from "../api";
+import { parseProjectNumber } from "./utils";
 
 const TIMEOUT = 10000;
 
@@ -15,16 +16,12 @@ export async function listTopVersions(
   versionCount: number,
   issueId?: string,
 ): Promise<string> {
+  const requestProjectId = parseProjectNumber(appId);
   try {
     const queryParams = new URLSearchParams();
     queryParams.set("page_size", `${versionCount}`);
     if (issueId) {
       queryParams.set("filter.issue.id", issueId);
-    }
-
-    const requestProjectId = parseProjectId(appId);
-    if (requestProjectId === undefined) {
-      throw new FirebaseError("Unable to get the projectId from the AppId.");
     }
 
     logger.debug(`[mcp][crashlytics] listTopVersions query paramaters: ${queryParams}`);
@@ -46,12 +43,4 @@ export async function listTopVersions(
       { original: err },
     );
   }
-}
-
-function parseProjectId(appId: string): string | undefined {
-  const appIdParts = appId.split(":");
-  if (appIdParts.length > 1) {
-    return appIdParts[1];
-  }
-  return undefined;
 }
